# executor: enhance memory tracking for IndexLookup